+++ src/main/kotlin/dag2/SimpleDiveCommander.kt
@@ -1,7 +1,7 @@
 package dag2
 
-open class SimpleDiveCommander(protected var depth: Int = 0, protected var horizontal: Int = 0) : CommandController {
-    override fun takeCommand(command: Command, num: Int) {
+open class SimpleDiveCommander(protected var depth: Long = 0, protected var horizontal: Long = 0) : CommandController {
+    override fun takeCommand(command: Command, num: Long) {
         when (command) {
             Command.FORWARD -> horizontal += num
             Command.UP -> depth -= num
@@ -9,7 +9,7 @@ open class SimpleDiveCommander(protected var depth: Int = 0, protected var horiz
         }
     }
 
-    override fun result(): Int {
+    override fun result(): Long {
         return depth * horizontal;
     }
 }
